Update tagUserNotifcation.php
This plugin was not working in WordPress 3.9.1. Here's how I got it working:

In v0.1 of this plugin, there's a file called:
tagUserNotifcation.php

On line 88, it's calling the option "notifyOnUserTag_emailTemplate" which does not exist by default, nor does the plugin add the option.

So on line 92, I delete the first part "if($template )" so it's shows this instead:
$template = file_get_contents(dirname(__FILE__) . '/defaultTemplate.php');

Then I deleted line 88: 
$template = get_option('notifyOnUserTag_emailTemplate');

Also, RegEx on line 23 didn't work with users names that contained . or -; So I modified the pattern to allow them.